Eliminated a pack of compiler warnings (gcc-4.6.1 amd64)

+++ dw/fltkui.cc
@@ -865,7 +865,7 @@ template <class I>
 {
    ::fltk::Item *item = new ::fltk::Item (name);
    item->set_flag (::fltk::RAW_LABEL);
-   item->user_data ((void *) index);
+   item->user_data ((void *)(long) index);
    return item;
 }
